rename pkg redissync -> statesync
